requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
asgiref==3.9.1
    # via
    #   -r requirements/base.txt
    #   django
click==8.2.1
    # via code-annotations
code-annotations==2.3.0
    # via -r requirements/test.in
coverage[toml]==7.10.2
    # via pytest-cov
    # via
    #   -c https:/raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-model-utils
django-model-utils==5.0.0
    # via -r requirements/base.txt
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
openedx-atlas==0.7.0
    # via -r requirements/base.txt
packaging==25.0
    # via pytest
pbr==6.1.1
    # via stevedore
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.2.1
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via code-annotations
text-unidecode==1.3
    # via python-slugify

# The following packages are considered to be unsafe in a requirements file:
# setuptools
